前端配置文件调整

+++ src/util/wxShare.js
@@ -1,68 +1,15 @@
-import { getJSSDK,payorders,xcpayorders } from '@/api/wx';
-import router from '../router'
-// 朋友圈
-// export default function wxShare() {
-//     if (!isWeixinBrowser()) {
-//         return;
-//     }
-//     loadShareSignature();
-//     wx.ready(function () {
-//         wx.getLocation({
-//             type: 'wgs84', // 默认为wgs84的gps坐标，如果要返回直接给openLocation用的火星坐标，可传入'gcj02'
-//             success: response => {
-//                 // alert('地理位置获取成功',JSON.stringify(response))
-//                 localStorage.setItem('latlon',JSON.stringify(response))
-//             },
-//             fail: err => {
-//                 alert('获取位置失败', JSON.stringify(err))
-//             },
-//             cancel: err => {
-//                 alert('用户拒绝授权获取地理位置', err)
-//             }
-//         })  
-//     });
-//     wx.error(p => {
-//         console.log(p)
-//     });
-// }
+import { getJSSDK} from '@/api/wx';
+import router from '@/router'
 const wxShare = () => {
     return new Promise((resolve, reject) => {
         if (!isWeixinBrowser()) {
             return;
         }
-        // let shareSignature = JSON.parse(sessionStorage.shareSignature);
-        // if(shareSignature){
-        //     setShareConfig(shareSignature);
-        //     resolve()
-        //     return
-        // }
-    let url = encodeURIComponent(window.location.origin + router.currentRoute.fullPath)
-    // if(sessionStorage.shareSignature){
-    //     let jssdk = JSON.parse(sessionStorage.shareSignature)
-    //     wx.config({
-    //         debug: true,
-    //         appId: 'wx2b58cb8bd7d7ceb1',
-    //         timestamp: jssdk.timestamp,
-    //         nonceStr: jssdk.nonceStr,
-    //         signature: jssdk.signature,
-    //         jsApiList: [
-    //             'checkJsApi',
-    //             'onMenuShareTimeline',
-    //             'onMenuShareAppMessage',
-    //             'getLocation',
-    //             'hideMenuItems',
-    //             'chooseWXPay']  
-    //     });
-    //     resolve()
-    // }else{
-        getJSSDK(url).then( data => {
-            if(data){
+        let url = encodeURIComponent(window.location.origin + router.currentRoute.fullPath)
+        getJSSDK(url).then(data => {
+            if (data) {
                 sessionStorage.shareSignature = JSON.stringify(data.data)
                 let userInfo = JSON.parse(localStorage.getItem('userInfo'))
-                // setShareConfig(data.data.data);
-                // alert(data.data.data)
-                debugger
-                console.log('http://api.welaipay.com/#/?inviteCode='+userInfo.invite_code)
                 wx.config({
                     debug: false,
                     appId: 'wx2b58cb8bd7d7ceb1',
@@ -75,10 +22,10 @@ const wxShare = () => {
                         'onMenuShareAppMessage',
                         'getLocation',
                         'hideMenuItems',
-                        'chooseWXPay']  
+                        'chooseWXPay']
                 });
-                wx.onMenuShareTimeline({ 
-                    link: 'http://api.welaipay.com/#/?inviteCode='+userInfo.invite_code,
+                wx.onMenuShareTimeline({
+                    link: 'http://api.welaipay.com/#/?inviteCode=' + userInfo.invite_code,
                     title: '刚发现一个神器，用它去加油能打9折，不需要充值',
                     desc: '直接去加油站支付油费就可以，我刚加一次油便宜了几十',
                     imgUrl: 'https://shengxin-static.oss-cn-shenzhen.aliyuncs.com/images/vehicle/life/img-share-thumb.png',
@@ -87,8 +34,8 @@ const wxShare = () => {
                         alert('分享成功')
                     }
                 })
-                wx.onMenuShareAppMessage({ 
-                    link: 'http://api.welaipay.com/#/?inviteCode='+userInfo.invite_code,
+                wx.onMenuShareAppMessage({
+                    link: 'http://api.welaipay.com/#/?inviteCode=' + userInfo.invite_code,
                     title: '刚发现一个神器，用它去加油能打9折，不需要充值',
                     desc: '直接去加油站支付油费就可以，我刚加一次油便宜了几十',
                     imgUrl: 'https://shengxin-static.oss-cn-shenzhen.aliyuncs.com/images/vehicle/life/img-share-thumb.png',
@@ -98,41 +45,19 @@ const wxShare = () => {
                     }
                 })
                 resolve()
-            }else if(data.data.code==401){
+            } else if (data && data.data.code == 401) {
                 localStorage.clear()
                 sessionStorage.clear()
                 window.location.reload()
-            }else{
+            } else {
                 alert(JSON.stringify(data))
             }
         })
-    // }
-    wx.error(p => {
+        // }
+        wx.error(p => {
             console.log(p)
         });
     })
-}    
-const loadShareSignature = () => {
-    
-    let url = window.location.origin + router.currentRoute.fullPath
-    // alert('没有config')
-    getJSSDK(url).then( data => {
-        // alert('获取到了config'+data.data.code)
-        console.log(111)
-        if(data.data.code==0){
-            sessionStorage.shareSignature = JSON.stringify(data.data.data)
-            setShareConfig(data.data.data);
-        }else if(data.data.code==401){
-            localStorage.clear()
-            sessionStorage.clear()
-            window.location.reload()
-        }else{
-            alert(JSON.stringify(data))
-        }
-        
-    })
-    
-    
 }
 
 function setShareConfig(shareSignature) {
@@ -148,23 +73,8 @@ function setShareConfig(shareSignature) {
             'onMenuShareAppMessage',
             'getLocation',
             'hideMenuItems',
-            'chooseWXPay']  
+            'chooseWXPay']
     });
-    // wx.ready(function () {
-        // wx.getLocation({
-        //     type: 'wgs84', // 默认为wgs84的gps坐标，如果要返回直接给openLocation用的火星坐标，可传入'gcj02'
-        //     success: response => {
-        //         // alert('地理位置获取成功',JSON.stringify(response))
-        //         localStorage.setItem('latlon',JSON.stringify(response))
-        //     },
-        //     fail: err => {
-        //         alert('获取位置失败1', JSON.stringify(err))
-        //     },
-        //     cancel: err => {
-        //         alert('用户拒绝授权获取地理位置', err)
-        //     }
-        // })  
-    // });
 }
 
 function isWeixinBrowser() {
